+++ test/raptor/test_cli.rb
@@ -34,6 +34,7 @@ module Raptor
       assert_nil options(cli)[:pid_file]
       assert_nil options(cli)[:stdout_file]
       assert_nil options(cli)[:stderr_file]
+      assert_nil options(cli)[:access_log_file]
     end
 
     def test_rackup_file_positional_argument
@@ -235,6 +236,12 @@ module Raptor
       cli = CLI.new(["--stderr-file", "/tmp/raptor.err"])
 
       assert_equal "/tmp/raptor.err", options(cli)[:stderr_file]
+    end
+
+    def test_access_log_file
+      cli = CLI.new(["--access-log-file", "/tmp/raptor.access.log"])
+
+      assert_equal "/tmp/raptor.access.log", options(cli)[:access_log_file]
     end
 
     def test_multiple_options_together